requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==23.1.0
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
discord.py==1.7.3
freetype-py==2.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.5.0
pipdeptree==2.7.1
pycairo==1.23.0
python-decouple==3.8
reportlab==4.0.0
requests==2.31.0
rlPyCairo==0.2.0
typing_extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
